Mgrate ABViewDocxBuilder to plugin
